requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
asyncpg==0.21.0
certifi==2020.06.20
discord==1.0.1
python-dotenv==0.15.0